+++ src/Aardvark.Rendering/Surfaces/Surface.fs
@@ -13,7 +13,7 @@ type DynamicSurface = EffectInputLayout * aval<Imperative.Module>
 [<RequireQualifiedAccess; ReferenceEquality>]
 type Surface =
     | Effect of effect: Effect
-    | Dynamic of compile: (EffectConfig -> DynamicSurface)
+    | Dynamic of compile: (IFramebufferSignature -> IndexedGeometryMode -> DynamicSurface)
     | Backend of surface: IBackendSurface
     | None
 
@@ -21,18 +21,19 @@ type Surface =
     static member FShadeSimple(effect: Effect) = Surface.Effect effect
 
     [<Obsolete("Use Surface.Dynamic instead.")>]
-    static member FShade(compile: EffectConfig -> DynamicSurface) = Surface.Dynamic compile
+    static member FShade(compile: IFramebufferSignature -> IndexedGeometryMode -> DynamicSurface) = Surface.Dynamic compile
 
 [<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
 module Surface =
 
     [<Sealed; AbstractClass>]
     type Converter() =
-        static member inline ToSurface(surface: Surface)                       = surface
-        static member inline ToSurface(effect: Effect)                         = Surface.Effect effect
-        static member inline ToSurface(effects: #seq<Effect>)                  = Surface.Effect <| FShade.Effect.compose effects
-        static member inline ToSurface(compile: EffectConfig -> DynamicSurface) = Surface.Dynamic compile
-        static member inline ToSurface(surface: IBackendSurface)               = Surface.Backend surface
+        static member inline ToSurface(surface: Surface)         = surface
+        static member inline ToSurface(effect: Effect)           = Surface.Effect effect
+        static member inline ToSurface(effects: #seq<Effect>)    = Surface.Effect <| FShade.Effect.compose effects
+        static member inline ToSurface(compile)                  = Surface.Dynamic compile
+        static member inline ToSurface(compile: Func<_, _, _>)   = Surface.Dynamic (fun s t -> compile.Invoke(s, t))
+        static member inline ToSurface(surface: IBackendSurface) = Surface.Backend surface
 
     let inline private toSurface (_ : ^Z) (data: ^T) =
         ((^Z or ^T) : (static member ToSurface : ^T -> Surface)(data))